Reorganize mobile components
