+++ packages/desktop-client/src/components/mobile/accounts/AccountDetails.jsx
@@ -1,19 +1,19 @@
 import React, { useState, useMemo } from 'react';
 
-import { useActions } from '../../hooks/useActions';
-import { SvgAdd } from '../../icons/v1';
-import { SvgSearchAlternate } from '../../icons/v2';
-import { theme } from '../../style';
-import { ButtonLink } from '../common/ButtonLink';
-import { InputWithContent } from '../common/InputWithContent';
-import { Label } from '../common/Label';
-import { View } from '../common/View';
-import { MobileBackButton } from '../MobileBackButton';
-import { Page } from '../Page';
-import { PullToRefresh } from '../responsive/PullToRefresh';
-import { CellValue } from '../spreadsheet/CellValue';
-import { useSheetValue } from '../spreadsheet/useSheetValue';
-import { TransactionList } from '../transactions/MobileTransaction';
+import { useActions } from '../../../hooks/useActions';
+import { SvgAdd } from '../../../icons/v1';
+import { SvgSearchAlternate } from '../../../icons/v2';
+import { theme } from '../../../style';
+import { ButtonLink } from '../../common/ButtonLink';
+import { InputWithContent } from '../../common/InputWithContent';
+import { Label } from '../../common/Label';
+import { View } from '../../common/View';
+import { MobileBackButton } from '../../MobileBackButton';
+import { Page } from '../../Page';
+import { CellValue } from '../../spreadsheet/CellValue';
+import { useSheetValue } from '../../spreadsheet/useSheetValue';
+import { PullToRefresh } from '../PullToRefresh';
+import { TransactionList } from '../transactions/TransactionList';
 
 function TransactionSearchInput({ accountName, onSearch }) {
   const [text, setText] = useState('');
